bare bones of an "editor" view and a "management" view - note that the initial page is now a button choice to select the view, but it needs work.
