+++ src/main/webui/src/ColourSchemeToggle.tsx
@@ -2,7 +2,7 @@ import {useMantineColorScheme, Switch, useMantineTheme} from '@mantine/core';
 import {IconSun, IconMoonStars} from '@tabler/icons-react';
 import { STROKE } from './constants.tsx';
 
-export function SwitchToggle() {
+export function ColourSchemeToggle() {
     const { colorScheme, toggleColorScheme } = useMantineColorScheme();
     // the colour gray used by the tools.
     const theme = useMantineTheme();
